[rename] ChangeCompanyStatusCommand로 네이밍 변경
