+++ company/src/main/java/com/hubEleven/company/application/command/ChangeCompanyStatusCommand.java
@@ -1,11 +1,10 @@
 package com.hubEleven.company.application.command;
 
 import com.hubEleven.company.domain.model.CompanyStatus;
-import com.hubEleven.company.domain.model.CompanyType;
 
 import java.util.UUID;
 
-public record ChangeStatusCompanyCommand(
+public record ChangeCompanyStatusCommand(
         UUID companyId,
         CompanyStatus status
 ) {
