+++ company/src/main/java/com/hubEleven/company/application/dto/response/CompanyResult.java
@@ -1,4 +1,4 @@
-package com.hubEleven.company.application.dto;
+package com.hubEleven.company.application.dto.response;
 
 import com.hubEleven.company.domain.model.Company;
 import com.hubEleven.company.domain.model.CompanyStatus;
@@ -6,7 +6,7 @@ import com.hubEleven.company.domain.model.CompanyType;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-public record CompanyDTO(
+public record CompanyResult(
 		UUID companyId,
 		UUID hubId,
 		String name,
@@ -16,8 +16,8 @@ public record CompanyDTO(
 		String address,
 		LocalDateTime createdAt,
 		LocalDateTime updatedAt) {
-	public static CompanyDTO from(Company company) {
-		return new CompanyDTO(
+	public static CompanyResult from(Company company) {
+		return new CompanyResult(
 				company.getCompanyId(),
 				company.getHubId(),
 				company.getName(),
